(2-14-24) 5:59 P.M. Made a simple Main Menu stand in
Pink Background Image with Yellow Title Text (FLOWERING). Two buttons, Play and Quit. On the gameobject MainMenu, there's a script with a number that can be changed to whatever scene number you want the players to go to. The quit button will kick the players out of the game. This is a stand in for the future assets to replace.